+++ src/Lib.Services/Logging/Odesli/OdesliServiceLogging.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.Logging;
 
-namespace MuzakBot.App.Logging.Odesli;
+namespace MuzakBot.Lib.Services.Logging.Odesli;
 
 /// <summary>
 /// Source generated logging methods for the <see cref="Services.OdesliService"/> class.
